Revert "User authentication jwt"
