+++ frontend/.eslintrc.json
@@ -1,40 +1,30 @@
 {
-  "env": {
-    "browser": true,
-    "es2021": true
-  },
-  "extends": [
-    "eslint:recommended",
-    "plugin:react/recommended",
-    "plugin:prettier/recommended"
-  ],
-  "parser": "@babel/eslint-parser",
-  "parserOptions": {
-    "ecmaFeatures": {
-      "jsx": true
+    "env": {
+        "browser": true,
+        "es2021": true
     },
-    "ecmaVersion": "latest",
-    "sourceType": "module",
-    "requireConfigFile": false,
-    "babelOptions": {
-      "presets": ["@babel/preset-react"]
-    }
-  },
-  "plugins": ["react"],
-  "rules": {
-    "react/react-in-jsx-scope": 0,
-    "prettier/prettier": [
-      "error",
-      {
-        "endOfLine": "auto"
-      }
+    "extends": [
+        "plugin:react/recommended",
+        "standard",
+        "plugin:prettier/recommended"
     ],
-    "react/prop-types": 0
-  },
-
-  "settings": {
-    "react": {
-      "version": "detect"
+    "parserOptions": {
+        "ecmaFeatures": {
+            "jsx": true
+        },
+        "ecmaVersion": "latest",
+        "sourceType": "module"
+    },
+    "plugins": [
+        "react"
+    ],
+    "rules": {
+        "react/react-in-jsx-scope": 0
+    },
+    
+    "settings": {
+        "react": {
+            "version": "detect"
+        }
     }
-  }
 }
