+++ frontend/README.old.md
@@ -1,3 +1,2 @@
 # Efficionado
-
-Personal Web App to Track Efficiency.
+Personal Web App to Track Efficiency.
