--- frontend/src/App.js
@@ -1,57 +1,0 @@
-import React, { useState, useEffect } from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import axios from "axios";
-import Header from "./components/layout/Header";
-import Home from "./components/pages/Home";
-import Register from "./components/auth/Register";
-import Login from "./components/auth/Login";
-import UserContext from "./context/userContext";
-import "./App.css";
-
-function App() {
-  const [userData, setUserData] = useState({
-    token: undefined,
-    user: undefined,
-  });
-
-  useEffect(() => {
-    const checkLoggedIn = async () => {
-      let token = localStorage.getItem("auth-token");
-      if (token === null) {
-        localStorage.setItem("auth-token", "");
-        token = "";
-      }
-      const tokenResponse = await axios.post(
-        "http://localhost:5000/users/tokenIsValid",
-        null,
-        { headers: { "x-auth-token": token } }
-      );
-      if (tokenResponse.data) {
-        const userRes = await axios.get("http://localhost:5000/users/", {
-          headers: { "x-auth-token": token },
-        });
-        setUserData({
-          token,
-          user: userRes.data,
-        });
-      }
-    };
-
-    checkLoggedIn();
-  }, []);
-
-  return (
-    <BrowserRouter>
-      <UserContext.Provider value={{ userData, setUserData }}>
-        <Header />
-        <Routes>
-          <Route exact path="/" component={Home} />
-          <Route path="/register" component={Register} />
-          <Route path="/login" component={Login} />
-        </Routes>
-      </UserContext.Provider>
-    </BrowserRouter>
-  );
-}
-
-export default App;
