+++ frontend/src/createCards.js
@@ -1,5 +1,6 @@
 import React from "react";
 import styled from "styled-components";
+import Task from "./task";
 import { Droppable } from "react-beautiful-dnd";
 
 const Container = styled.div`
